requirements.txt
opencv-python
pillow
streamlit
matplotlib
ultralytics
pandas
plotly